Optimizations
A previous change that improved some function signatures by storing
extra type information decreased performance by 4.5x. This was due to
also generating many function overloads that would never actually end up
in the output file. This is fixed by adding a check before starting to
generate the functions, and exiting early if we can determine that the
function will never be written to any file.